Rename the account audit logs file.
